+++ Changelog.md
@@ -1,5 +1,6 @@
 --------------------------Changelog for v1.9.12 of SecurityCraft--------------------------
 
+- API: New method IPasscodeProtected#setPasscodeInAdjacentBlock for updating an adjacent block (e.g. the second half of a chest) with the passcode when setting it
 - Fix: Security Camera entities are sent to every player instead of only the player mounting the camera
 
 --------------------------Changelog for v1.9.11 of SecurityCraft--------------------------